+++ mipac/actions/mute.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+
 from typing import TYPE_CHECKING, AsyncIterable
 
 from mipac.abstract.action import AbstractAction