+++ mipac/manager/__init__.py
@@ -8,10 +8,10 @@ from .client import *
 from .drive import *
 from .favorite import *
 from .follow import *
+from .mute import *
 from .my import *
 from .note import *
 from .page import *
 from .poll import *
 from .reaction import *
 from .user import *
-from .mute import *