+++ mipac/models/__init__.py
@@ -5,9 +5,9 @@ from .emoji import *
 from .follow import *
 from .instance import *
 from .lite import *
+from .mute import *
 from .note import *
 from .notification import *
 from .poll import *
 from .reaction import *
 from .user import *
-from .mute import *