+++ mipac/types/__init__.py
@@ -7,10 +7,10 @@ from .emoji import *
 from .endpoints import *
 from .follow import *
 from .instance import *
+from .mute import *
 from .note import *
 from .notification import *
 from .page import *
 from .poll import *
 from .reaction import *
 from .user import *
-from .mute import *