+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ## Added
+
+- ✨ added `un_renote` method at `ClientNoteActions` class.
 - ✨ added `get_children` method at `ClientNoteActions` class.
 - ✨ added `invalidate` method at `FollowActions` class.
 - ✨ added `cancel` method at `FollowRequestActions` class.